fix: build watch mode not watching
